Move to /.../[hexId]/update URL
