+++ src/pages/manage-my-registrations/[hexId]/update/details.tsx
