+++ src/pages/manage-my-registrations/[hexId]/update/use/more-details.tsx
